% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exp_r2.R
\name{exp_r2}
\alias{exp_r2}
\title{Expected r2 between sMLH and inbreeding level (f)}
\usage{
exp_r2(genotypes, subsets = NULL, nboot = 100, type = c("msats", "snps"))
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and loci in columns,
containing genotypes coded as 0 (homozygote) and 1 (heterozygote)}

\item{subsets}{a vector specifying the sizes of subsets to draw. For a subset of 20 markers, subsets = c(2, 5, 10, 15, 20) could
be a reasonable choice. The minimum subset size is 2 and the maximum is the number of markers in the data.}

\item{nboot}{number re-draws per subset size.}

\item{type}{specifies g2 formula to take. Type "snps" for large datasets and "msats" for smaller datasets.}
}
\value{
\item{call}{function call.}
\item{exp_r2_full}{expected r2 between inbreeding and sMLH for the full dataset}
\item{exp_r2_res}{expected r2 for each randomly subsetted dataset}
\item{summary_exp_r2}{r2 mean and sd for each number of subsetted loci}
\item{nobs}{number of observations}
\item{nloc}{number of markers}
}
\description{
Expected r2 between sMLH and inbreeding level (f)
}
\examples{
data(mouse_msats)
genotypes <- convert_raw(mouse_msats, miss = NA)
(out <- exp_r2(genotypes, subsets = c(2,4,6,8,10,12), nboot = 1000, type = "msats"))
plot(out)
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Szulkin, M., Bierne, N., & David, P. (2010). HETEROZYGOSITY-FITNESS CORRELATIONS: A TIME FOR REAPPRAISAL.
Evolution, 64(5), 1202-1217.
}

