% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/g2_microsats.R
\name{g2_microsats}
\alias{g2_microsats}
\title{Estimating g2 from microsatellite data}
\usage{
g2_microsats(genotypes, nperm = 0, nboot = 0, CI = 0.95)
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and loci in columns,
containing genotypes coded as 0 (homozygote) and 1 (heterozygote)}

\item{nperm}{Number of permutations for testing the hypothesis that the empirical g2-value is higher than by chance}

\item{nboot}{Number of bootstraps for estimating a confidence interval}

\item{CI}{Confidence interval (default to 0.95)}
}
\value{
\item{call}{Model call.}
\item{g2}{g2 value}
\item{p_val}{p value from permutation test}
\item{g2_permut}{g2 values from permuted genotypes}
\item{g2_boot}{g2 values from bootstrap samples}
\item{CI_boot}{confidence interval from bootstraps}
\item{se_boot}{standard error of g2 from bootstraps}
\item{nobs}{number of observations}
\item{nloc}{number of markers}
}
\description{
Estimating g2 from microsatellite data
}
\examples{
data(seal_microsats)
# tranform raw genotypes into 0/1 format
genotypes <- convert_raw(seal_microsats, NAval = NA)
(g2_seals <- g2_microsats(genotypes, nperm = 100, nboot = 100, CI = 0.95))
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com) &
        Mareike Esser (messer@techfak.uni-bielefeld.de)
}
\references{
David, P., Pujol, B., Viard, F., Castella, V. and Goudet, J. (2007),
Reliable selfing rate estimates from imperfect population genetic data. Molecular Ecology,
16: 2474
}

