% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/g2_microsats_new.R
\name{g2_microsats_new}
\alias{g2_microsats_new}
\title{estimating second-order heterozygosity from genotype data}
\usage{
g2_microsats_new(genotypes, nperm = 0, nboot = 0, CI = 0.95)
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and loci in columns,
containing genotypes coded as 0 (homozygote) and 1 (heterozygote).}

\item{missing}{Value of missing data. -1 if data.frame is output from \link{convert_raw}.}
}
\value{
\item{call}{Model call.}
\item{g2}{g2 value}
\item{p_val}{p value from permutation test}
\item{g2_permut}{g2 values from permuted genotypes}
\item{g2_boot}{g2 values from bootstrap samples}
\item{CI_boot}{confidence interval from bootstraps}
\item{se_boot}{standard error of g2 from bootstraps}
\item{nobs}{number of observations}
\item{nloc}{number of markers}
}
\description{
estimating second-order heterozygosity from genotype data
}
\examples{
data(seal_microsats)
# tranform raw genotypes into 0/1 format
genotypes <- convert_raw(seal_microsats, NAval = NA)
(g2_val <- g2_microsats(genotypes, nperm = 100, nboot = 100, CI = 0.95))
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com) &
        Mareike Esser (messer@techfak.uni-bielefeld.de)
}
\references{
DAVID, P., PUJOL, B., VIARD, F., CASTELLA, V. and GOUDET, J. (2007),
Reliable selfing rate estimates from imperfect population genetic data. Molecular Ecology,
16: 2474–2487. doi: 10.1111/j.1365-294X.2007.03330.x
}

