% Generated by roxygen2 (4.0.2): do not edit by hand
\name{g2_snps}
\alias{g2_snps}
\title{Estimating g2 from SNP data}
\usage{
g2_snps(genotypes)
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and loci in columns,
       containing genotypes coded as 0 (homozygote) or 1 (heterozygote)}
}
\value{
g2 value
}
\description{
Estimating g2 from SNP data
}
\details{
Calculates g2 from big SNP datasets. Use convert_raw to convert raw genotypes (with 2 columns per locus) into
         the required format
}
\examples{
# load SNP genotypes, already in 0 (homozygous), 1(heterozygous) format.

data(mice_snp_genotypes)
g2_val <- g2_snps(genotypes)
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com) &
        Mareike Esser (messer@techfak.uni-bielefeld.de)
}
\references{
Hoffman, J.I., Simpson, F., David, P., Rijks, J.M., Kuiken, T., Thorne, M.A.S., Lacey, R.C. & Dasmahapatra, K.K. (2014) High-throughput sequencing reveals inbreeding depression in a natural population.
Proceedings of the National Academy of Sciences of the United States of America, 111: 3775-3780. Doi: 10.1073/pnas.1318945111
}

