% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HHC.R
\name{HHC}
\alias{HHC}
\title{Calculates heterzygosity-heterozygosity correlations with
standardized multilocus heterozygosities (sMLH)}
\usage{
HHC(genotypes, iter = 100)
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and loci in columns,
       containing genotypes coded as 0 (homozygote) and 1 (heterozygote)}
}
\value{
Vector of het-het correlations
}
\description{
Calculates heterzygosity-heterozygosity correlations with
standardized multilocus heterozygosities (sMLH)
}
\examples{
data(mice_snp_genotypes)
result <- HHC(mice_snp_genotypes, iter = 10)
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Balloux, F., Amos, W., & Coulson, T. (2004). Does heterozygosity estimate inbreeding
in real populations?. Molecular Ecology, 13(10), 3021-3031.
}

