% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mice_snp_genotypes.R
\name{mice_snp_genotypes}
\alias{mice_snp_genotypes}
\title{Mice SNP data}
\format{A data frame with 36 observations on 13199 loci.}
\description{
Mice snp data in 0 (homozygous), 1(heterzygous), -1 (missing) format.
Each individual is in a row and each locus in a column.
}
\references{
Hoffman, J.I., Simpson, F., David, P., Rijks, J.M., Kuiken, T., Thorne, M.A.S., Lacey, R.C. & Dasmahapatra, K.K. (2014) High-throughput sequencing reveals inbreeding depression in a natural population.
Proceedings of the National Academy of Sciences of the United States of America, 111: 3775-3780. Doi: 10.1073/pnas.1318945111
}
\keyword{datasets}

