% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.g2.R
\name{plot.g2}
\alias{plot.g2}
\title{Plot a g2 object
Plots the distribution of g2 estimates from bootstrapping.}
\usage{
\method{plot}{g2}(x, main = "g2 bootstrap estimates", xlab = "g2",
  ylab = "counts", ...)
}
\arguments{
\item{x}{A g2 object.}

\item{\dots}{Additional arguments to the hist() function.}
}
\description{
Plot a g2 object
Plots the distribution of g2 estimates from bootstrapping.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\seealso{
\link{g2_snps}, \link{g2_microsats}
}

