% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.inbreed.R
\name{plot.inbreed}
\alias{plot.inbreed}
\title{Plot a g2 object
Plots the distribution of g2 estimates from bootstrapping.}
\usage{
\method{plot}{inbreed}(x, main = "g2 bootstrap estimates", xlab = "g2",
  ylab = "counts", ...)
}
\arguments{
\item{x}{An g2 object.}

\item{main}{Plot title}

\item{xlab}{X-axis label}

\item{ylab}{y-axis label}

\item{\dots}{Additional arguments to the hist() function.}
}
\description{
Plot a g2 object
Plots the distribution of g2 estimates from bootstrapping.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\seealso{
\link{g2_snps}, \link{g2_microsats}
}

