% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.g2.R
\name{print.g2}
\alias{print.g2}
\title{Print a rpt object}
\usage{
\method{print}{g2}(x, ...)
}
\arguments{
\item{x}{A g2 object return from one of the g2 functions.}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Displays the results a rpt object (i.e. the result of a rpt function call) an a nice form.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\seealso{
\link{g2_snps}, \link{g2_microsats}, \link{plot}
}

