% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.inbreed.R
\name{print.inbreed}
\alias{print.inbreed}
\title{Print an inbreed object}
\usage{
\method{print}{inbreed}(x, ...)
}
\arguments{
\item{x}{An \code{inbreed} object from one of the \code{inbreedR} functions.}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Displays the results a \code{inbreed} object.
}
\seealso{
\link{g2_snps}, \link{g2_microsats}, \link{plot}
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}
