% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seal_microsats.R
\name{seal_microsats}
\alias{seal_microsats}
\title{Antarctic fur eal microsatellite data}
\format{A data frame with 82 individuals and 41 microsatellite markers.}
\description{
Dataset with each microsatellite locus on two adjecent columns (one per allel)
}
\references{
Stoffel, M.A, Caspers, B.A., Forcada, J., Giannakara, A., Baier, M.C., Eberhart-Phillips, L.J., Mueller, C. & Hoffman, J.I. (in review)
}
\keyword{datasets}

