% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seal_msats.R
\name{seal_msats}
\alias{seal_msats}
\title{Harbour seal microsatellite data}
\format{A data frame with 60 individuals and 27 microsatellite markers.}
\description{
Dataset with each microsatellite locus in two adjecent columns (one per allel)
}
\references{
Hoffman, J.I., Simpson, F., David, P., Rijks, J.M., Kuiken, T., Thorne, M.A.S., Lacey, R.C. & Dasmahapatra, K.K. (2014) High-throughput sequencing reveals inbreeding depression in a natural population.
Proceedings of the National Academy of Sciences of the United States of America, 111: 3775-3780. Doi: 10.1073/pnas.1318945111
}
\keyword{datasets}

