% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seal_snps.R
\name{seal_snps}
\alias{seal_snps}
\title{Seal SNP data}
\format{A data frame with 60 observations on 14596 loci.}
\description{
Seal snp data in 0 (homozygous), 1(heterzygous), -1 (missing) format.
Each row represents an individual and each column is a locus.
}
\details{
Original data from Hoffman et al. (2013) was randomly subsetted for half the loci for computational reasons.
}
\references{
Hoffman, J.I., Simpson, F., David, P., Rijks, J.M., Kuiken, T., Thorne, M.A.S., Lacey, R.C. & Dasmahapatra, K.K. (2014) High-throughput sequencing reveals inbreeding depression in a natural population.
Proceedings of the National Academy of Sciences of the United States of America, 111: 3775-3780. Doi: 10.1073/pnas.1318945111
}
\keyword{datasets}

