% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate_g2.R
\name{simulate_g2}
\alias{simulate_g2}
\title{Simulate g2}
\usage{
simulate_g2(n_ind = NULL, subsets = NULL, reps = 100, genotypes = NULL,
  mean_loc_MLH = NULL, sd_loc_MLH = NULL)
}
\arguments{
\item{n_ind}{number of individuals to sample from the population}

\item{subsets}{a vector specifying the sizes of marker-subsets to draw. For a subset of 20 markers, subsets = c(2, 5, 10, 15, 20) could
be a reasonable choice. The minimum subset size is 2 and the maximum is the number of markers in the data}

\item{reps}{number of resampling repetitions}

\item{genotypes}{optional: provide genotypes in \code{inbreedR} format to estimate the empricial multilocus heterozygosity
(MLH) across loci and use these estimates to simulate data from the same distribution}

\item{mean_loc_MLH}{mean multilocus heterozygosity across loci. Should range between 0-1. This will be automatically
calculated from the empirical data when genotypes in inbreedR format are given}

\item{sd_loc_MLH}{standard deviation of multilocus heterozygosity across loci.}
}
\description{
Simulates g2 values based on different marker numbers.
}
\details{
The \code{simulate_g2} function can be used to explore the confidence of a g2 estimate calculated
         from different sized marker sets. The simulation assumes (1) unlinked loci,
         (2) equal allele frequencies among all loci with expected heterozygosity of 0.5 and
         (3) random mating. The mean and standard deviation of genome-wide expected
         heterozygosity can either be specified with the \code{mean_loc_MLH} and \code{sd_loc_MLH}
         arguments or will be calculated from emprical genotypes when these are specified in
         the \code{genotypes} argument.
}
\author{
Marty Kardos &
         Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}

