% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_r2_hf.R
\name{simulate_r2_hf}
\alias{simulate_r2_hf}
\title{Calculates the expected squared correlation between heteorzygosity and inbreeding for simulated marker sets}
\usage{
simulate_r2_hf(n_ind = NULL, subsets = NULL, reps = 100,
  type = c("msats", "snps"), genotypes = NULL, mean_MLH = NULL,
  sd_MLH = NULL, CI = 0.95)
}
\arguments{
\item{n_ind}{number of individuals to sample from the population}

\item{subsets}{a vector specifying the sizes of marker-subsets to draw. Specifying  \code{subsets = c(2, 5, 10, 15, 20)} 
woud draw marker sets of 2 to 20 markers. The minimum number of markers to calculate g2 is 2.}

\item{reps}{number of resampling repetitions}

\item{type}{specifies g2 formula. Type "snps" for large datasets and "msats" for smaller datasets.}

\item{genotypes}{optional: provide genotypes in \code{inbreedR} format to estimate the empirical multilocus heterozygosity
(MLH) and use these estimates to simulate data from the same distribution}

\item{mean_MLH}{mean multilocus heterozygosity. Should range between 0-1. This will be automatically
calculated from the empirical data when genotypes in \code{inbreedR} format are given.}

\item{sd_MLH}{standard deviation of multilocus heterozygosity. This will be automatically
calculated from the empirical data when genotypes in \code{inbreedR} format are given.}

\item{CI}{Confidence intervals to calculate (default to 0.95)}
}
\description{
This function can be used to simulate genotype data, draw random subsamples and calculate the
respective expected squared correlations between heterozygosity and fitness (r2(h,f)). 
Every subsample of markers is drawn independently to give insights
into the magnitude and variation of r2(h,f) values calculated from a given number of markers and individuals. 
If empirical genotypes are given, the function will calculate the mean and sd of multilocus heterozygosity (MLH).
Each indiviudals' MLH and the zygosity of simulated loci are then randomly drawn from a normal distribution with mean = mean(MLH)
and sd = sd(MLH).
}
\details{
The \code{simulate_g2} function can be used to explore the confidence of g2 estimates calculated
         from marker sets of different sizes. Every new locus set is drawn independently.
         The simulation assumes (1) unlinked loci, (2) equal allele frequencies among all loci 
         with expected heterozygosity of 0.5 and (3) random mating. The mean and standard 
         deviation of genome-wide expected heterozygosity can either be specified with 
         the \code{mean_MLH} and \code{sd_MLH} arguments or will be calculated 
         from emprical genotypes when these are specified in the \code{genotypes} argument.
}
\examples{
data(mouse_msats)
genotypes <- convert_raw(mouse_msats)
sim_r2 <- simulate_r2_hf(n_ind = 10, subsets = c(2,4,6), reps = 100, genotypes = genotypes)
plot(sim_r2)
}
\author{
Marty Kardos (marty.kardos@ebc.uu.se) &
         Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}

