% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sMLH.R
\name{sMLH}
\alias{sMLH}
\title{Calculate standardized multilocus heterozygosity (sMLH)}
\usage{
sMLH(genotypes)
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and nloci in columns,
       containing genotypes coded as 0 (homozygote) and 1 (heterozygote)}
}
\value{
Vector of individual standardized multilocus heterozygosities
}
\description{
sMLH is defined as the total number of heterozygous loci in an individual divided
by the sum of average observed heterozygosities in the population over the
subset of loci successfully typed in the focal individual
}
\examples{
data(seal_msats)
genotypes <- convert_raw(seal_msats, miss = NA)
het <- sMLH(genotypes)
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Coltman, D. W., Pilkington, J. G., Smith, J. A., & Pemberton, J. M. (1999).
Parasite-mediated selection against inbred Soay sheep in a free-living,
island population. Evolution, 1259-1267.
}

