% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/var_in_het.R
\name{var_in_het}
\alias{var_in_het}
\title{Calculate standardized multilocus heterozygosity (sMLH) as well as a distribution
of sMLH under the hypothesis of no inbreeding (g2 = 0).}
\usage{
var_in_het(genotypes, nperm = 100)
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and loci in columns,
containing genotypes coded as 0 (homozygote) and 1 (heterozygote)}

\item{nperm}{Number of permutations in order to get the distribution of sMLH values
when there is no inbreeding.}
}
\value{
\item{sMLH_perm_mean}{Means of sMLH´s of permuted data.}
\item{sMLH_perm_mean}{Variances of sMLH´s of permuted data.}
\item{emp_sMLH}{empirical sMLH}
}
\description{
This function has the purpose of comparing the distribution of sMLH
in the sample to a theoretical distribution of sMLH when there is no inbreeding.
Distributions of sMLH under the hypothesis g2 = 0 are obtained by shuffling
genotypes at each locus randomly across individual.s
}
\examples{
data(seal_microsats)
# tranform raw genotypes into 0/1 format
genotypes <- convert_raw(seal_microsats, NAval = NA)
het <- sMLH(genotypes)
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Coltman, D. W., Pilkington, J. G., Smith, J. A., & Pemberton, J. M. (1999).
Parasite-mediated selection against inbred Soay sheep in a free-living,
island population. Evolution, 1259-1267.
}

