% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Checks the data for consistency with the inbreedR working format.}
\usage{
check_data(genotypes, num_ind = NULL, num_loci = NULL)
}
\arguments{
\item{genotypes}{data frame with individuals in rows and one locus per column,
containing genotypes coded as 0 (homozygote) and 1 (heterozygote)}

\item{num_ind}{Number of individuals}

\item{num_loci}{Number of loci / markers}
}
\value{
TRUE if the data format is correct, error message if any test failed
}
\description{
inbreedR requires each column to be a locus and one row per individual.
Heterozygosity at a given locus should be coded as 1 and Homozygosity as 0. Missing values
can be coded as NA or any number (other than 0 or 1).
}
\details{
Checks that (1) the genotype data just contains 3 elements, which is 0 for homozygote,
         1 for heterozygote and either NA or any number (except 0 and 1) for missing data, (2) the number
         of individuals corresponds to the number of rows and the number of loci corresponds to the
         number of columns, (3) the data type is numeric.
     .
}
\examples{
data(mouse_msats)
# tranform raw genotypes into 0/1 format
genotypes <- convert_raw(mouse_msats, miss = NA)
# check data
check_data(genotypes, num_ind = 36, num_loci = 12)
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}

