% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_raw.R
\name{convert_raw}
\alias{convert_raw}
\title{Genotype format converter}
\usage{
convert_raw(genotypes, NAval = -99)
}
\arguments{
\item{genotypes}{Raw genotype data frame or matrix. Has individuals in rows and each locus in two columns}

\item{NAval}{The prespecified value for missing elements in the data frame. Will be converted to -1 in the output.}
}
\value{
Data.frame object with 0 (homozygote), 1 (heterozygote) and -1 (NA on at least one locus).
        Each locus is thus represented by one column.
}
\description{
Turns raw genotype data into 0 (homozygote), 1 (heterozygote) and -1 (NA on at least one locus).
A raw genotype matrix has individuals in rows and loci in columns, while each locus has two adjacent.
Type \emph{data(seal_microsats)} for an example data frame.
}
\examples{
# Seal microsatellite data with missing values specified with -99
data(seal_microsats)
genotypes <- convert_raw(seal_microsats, NAval = NA)
head(genotypes)
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}

