% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exp_r2.R
\name{exp_r2}
\alias{exp_r2}
\title{Expected r2 between sMLH and inbreeding level}
\usage{
exp_r2(genotypes)
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and loci in columns,
       containing genotypes coded as 0 (homozygote) and 1 (heterozygote)}
}
\description{
Expected r2 between sMLH and inbreeding level
}
\examples{
data(mice_snp_genotypes)
result <- exp_r2(mice_snp_genotypes)
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Szulkin, M., Bierne, N., & David, P. (2010). HETEROZYGOSITY-FITNESS CORRELATIONS: A TIME FOR REAPPRAISAL.
Evolution, 64(5), 1202-1217.
}

