% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exp_r2.R
\name{exp_r2}
\alias{exp_r2}
\title{Expected r2 between sMLH and inbreeding level}
\usage{
exp_r2(genotypes, steps = 10, nboot = 100, type = c("msats", "snps"))
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and loci in columns,
containing genotypes coded as 0 (homozygote) and 1 (heterozygote)}

\item{steps}{specifies number of subsets for increasing marker number}

\item{nboot}{number of bootstraps per step}

\item{type}{specifies g2 formula to take. Type "snps" for large datasets and "msats" for smaller datasets.}
}
\value{
\item{call}{function call.}
\item{exp_r2_res}{expected r2 for each randomly subsetted dataset}
\item{summary_exp_r2}{r2 mean and sd for each number of subsetted loci}
\item{nobs}{number of observations}
\item{nloc}{number of markers}
}
\description{
Currently just working for microsat data!
}
\examples{
data(seal_snps)
(out <- exp_r2(seal_snps, steps = 5, nboot = 5, type = "snps"))
plot(out)
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Szulkin, M., Bierne, N., & David, P. (2010). HETEROZYGOSITY-FITNESS CORRELATIONS: A TIME FOR REAPPRAISAL.
Evolution, 64(5), 1202-1217.
}

