% Generated by roxygen2 (4.0.2): do not edit by hand
\name{g2}
\alias{g2}
\title{estimating second-order heterozygosity from genotype data}
\usage{
g2(genotypes)
}
\arguments{
\item{genotypes}{genotype matrix with each locus in two adjacent columns}
}
\value{
g2
}
\description{
estimating second-order heterozygosity from genotype data
}

