% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/g2.R
\docType{package}
\name{g2}
\alias{g2}
\alias{g2-package}
\title{g2: Estimating identity disequilibrium for microsatellite and snp data}
\description{
\link{g2_microsats}
\link{g2_snps}
\link{convert_raw}
\link{plot.g2}
}
\note{
Currently there is are a number of functions that use the two arguments \code{y} and \code{groups} to
estimate standard repeatabilities (see e.g. function \link{rpt}). There is another group of
functions (e.g. \link{rpt.adj}) that uses the agruments \code{formula} and \code{grname}, which allows
estimating adjusted repeatabilities (controlling for fixed effects) and the estimation of multiple variance
components simulatneously (multiple random effects). In the long run, the two groups of functions will be merged
and will use the more flexible fomula arguments. So far, adjusted repeatabilities are only implemented for
Gaussian data using REML estimation.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com), Mareike Esser (messer@uni-bielefeld.de)
}
\references{
DAVID, P., PUJOL, B., VIARD, F., CASTELLA, V. and GOUDET, J. (2007),
Reliable selfing rate estimates from imperfect population genetic data. Molecular Ecology,
16: 2474–2487. doi: 10.1111/j.1365-294X.2007.03330.x

Hoffman, J.I., Simpson, F., David, P., Rijks, J.M., Kuiken, T., Thorne, M.A.S., Lacey, R.C. & Dasmahapatra, K.K. (2014) High-throughput sequencing reveals inbreeding depression in a natural population.
Proceedings of the National Academy of Sciences of the United States of America, 111: 3775-3780. Doi: 10.1073/pnas.1318945111
}

