% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/g2_snps.R
\name{g2_snps}
\alias{g2_snps}
\title{Estimating g2 from SNP data}
\usage{
g2_snps(genotypes, nperm = 0, nboot = 0, CI = 0.95)
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and loci in columns, containing genotypes coded as 0 (homozygote) or 1 (heterozygote)}

\item{nperm}{number or permutations for calculating the p-value}

\item{nboot}{number of bootstraps for CI}

\item{CI}{confidence interval (default to 0.95)}
}
\value{
\item{call}{Model call.}
\item{g2}{g2 value}
\item{p_val}{p value from permutation test}
\item{g2_permut}{g2 values from permuted genotypes}
\item{g2_boot}{g2 values from bootstrap samples}
\item{CI_boot}{confidence interval from bootstraps}
\item{se_boot}{standard error of g2 from bootstraps}
\item{nobs}{number of observations}
\item{nloc}{number of markers}
}
\description{
Estimating g2 from SNP data
}
\details{
Calculates g2 from big SNP datasets. Use convert_raw to convert raw genotypes (with 2 columns per locus) into
         the required format
}
\examples{
# load SNP genotypes in 0 (homozygous), 1(heterozygous) format.

data(mice_snp_genotypes)
(g2_mice <- g2_snps(mice_snp_genotypes, nperm = 10, nboot = 10, CI = 0.95))
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com) &
        Mareike Esser (messer@techfak.uni-bielefeld.de)
}
\references{
Hoffman, J.I., Simpson, F., David, P., Rijks, J.M., Kuiken, T., Thorne, M.A.S., Lacey, R.C. & Dasmahapatra, K.K. (2014) High-throughput sequencing reveals inbreeding depression in a natural population.
Proceedings of the National Academy of Sciences of the United States of America, 111: 3775-3780. Doi: 10.1073/pnas.1318945111
}

