% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inbreedR.R
\docType{package}
\name{inbreedR}
\alias{inbreedR}
\alias{inbreedR-package}
\title{inbreedR: Workflows for analysing variance in inbreeding based on SNP or microsatellite markers.}
\description{
\link{g2_microsats}
\link{g2_snps}
\link{convert_raw}
\link{plot.g2}
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com), Mareike Esser (messer@uni-bielefeld.de)
}
\references{
DAVID, P., PUJOL, B., VIARD, F., CASTELLA, V. and GOUDET, J. (2007),
Reliable selfing rate estimates from imperfect population genetic data. Molecular Ecology,
16: 2474

Hoffman, J.I., Simpson, F., David, P., Rijks, J.M., Kuiken, T., Thorne, M.A.S., Lacey, R.C. & Dasmahapatra, K.K. (2014) High-throughput sequencing reveals inbreeding depression in a natural population.
Proceedings of the National Academy of Sciences of the United States of America, 111: 3775-3780.
}

