% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.inbreed.R
\name{plot.inbreed}
\alias{plot.inbreed}
\title{Plot an inbreed object}
\usage{
\method{plot}{inbreed}(x, plottype = c("boxplot", "histogram"), ...)
}
\arguments{
\item{x}{An \code{inbreed} object.}

\item{plottype}{deprecated. "boxplot" or "histogram" to plot the output of r2_hf() and to show either
the boxplots through resampling of loci or the histogram from the bootstrapping of 
r2 over individuals.}

\item{\dots}{Additional arguments to the hist() function for the g2 and HHC functions. 
Additional arguments to the boxplot() function for plotting the result of the r2_hf() function.}
}
\description{
Plot an inbreed object
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\seealso{
\link{g2_snps}, \link{g2_microsats}
}

