% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.inbreed.R
\name{plot.inbreed}
\alias{plot.inbreed}
\title{Plot an inbreed object}
\usage{
\method{plot}{inbreed}(x, ...)
}
\arguments{
\item{x}{An inbreed object.}

\item{\dots}{Additional arguments to the hist() function.}
}
\description{
Plot an inbreed object
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\seealso{
\link{g2_snps}, \link{g2_microsats}
}

