% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.inbreed.R
\name{print.inbreed}
\alias{print.inbreed}
\title{Print an inbreed object}
\usage{
\method{print}{inbreed}(x, ...)
}
\arguments{
\item{x}{An inbreed object from one of the inbreedR functions.}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Displays the results a inbreed object.
}
\author{
Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\seealso{
\link{g2_snps}, \link{g2_microsats}, \link{plot}
}

