% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/seal_microsats.R
\name{seal_microsats}
\alias{seal_microsats}
\title{Seal Microsatellite data}
\format{A data frame with 82 obvervations of 41 diploid microsatellite markers.}
\description{
Dataset with each microsatellite locus on two adjecent columns (one per allel)
}
\references{
Hoffman, J.I., Simpson, F., David, P., Rijks, J.M., Kuiken, T., Thorne, M.A.S., Lacey, R.C. & Dasmahapatra, K.K. (2014) High-throughput sequencing reveals inbreeding depression in a natural population.
Proceedings of the National Academy of Sciences of the United States of America, 111: 3775-3780. Doi: 10.1073/pnas.1318945111
}
\keyword{datasets}

