% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sMLH.R
\name{sMLH}
\alias{sMLH}
\title{Calculate standardized multinlocus heterozygosity (sMLH)}
\usage{
sMLH(genotypes)
}
\arguments{
\item{genotypes}{data.frame with individuals in rows and nloci in columns,
       containing genotypes coded as 0 (homozygote) and 1 (heterozygote)}
}
\value{
Vector of individual standardized multinlocus heterozygosities
}
\description{
sMLH is defined as total number of heterozygous nloci in an individual divided
by the sum of average observed heterozygosities in the population over the
subset of nloci successfully typed in the focal individual
}
\examples{
data(mice_snp_genotypes)
het <- sMLH(mice_snp_genotypes)
}
\author{
Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Coltman, D. W., Pilkington, J. G., Smith, J. A., & Pemberton, J. M. (1999).
Parasite-mediated selection against inbred Soay sheep in a free-living,
island population. Evolution, 1259-1267.
}

